requirements.txt
haystack-ai>=2.17.0
datasets>=4.0.0
neo4j>=5.0.0
py2neo>=2021.2.3
networkx>=3.0
matplotlib>=3.5.0
plotly>=5.0.0
flask>=2.0.0
pandas>=1.3.0
